Fix failing tests
* Use of immutable map in MessageLoggingAspect was throwing an
  exception and causing context to fail.
* Renamed package to appointments.subdelta.
