+++ src/itest/java/uk/gov/companieshouse/appointments/subdelta/Runner.java
@@ -1,0 +1,15 @@
+package uk.gov.companieshouse.appointments.subdelta;
+
+import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
+
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectClasspathResource;
+import org.junit.platform.suite.api.Suite;
+
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("features")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "uk.gov.companieshouse")
+public class Runner {
+}
