+++ src/main/java/uk/gov/companieshouse/appointments/subdelta/NullService.java
@@ -1,0 +1,15 @@
+package uk.gov.companieshouse.appointments.subdelta;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * The default service.
+ */
+@Component
+class NullService implements Service {
+
+    @Override
+    public void processMessage(ServiceParameters parameters) {
+        throw new NonRetryableException("Unable to handle message");
+    }
+}
