+++ src/main/java/uk/gov/companieshouse/appointments/subdelta/RetryableException.java
@@ -1,0 +1,15 @@
+package uk.gov.companieshouse.appointments.subdelta;
+
+/**
+ * A recoverable exception has occurred e.g. due to a service that is temporarily unavailable.
+ */
+public class RetryableException extends RuntimeException {
+
+    public RetryableException(String message) {
+        super(message);
+    }
+
+    public RetryableException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
