+++ src/main/java/uk/gov/companieshouse/appointments/subdelta/Service.java
@@ -1,0 +1,14 @@
+package uk.gov.companieshouse.appointments.subdelta;
+
+/**
+ * Processes an incoming message.
+ */
+public interface Service {
+
+    /**
+     * Processes an incoming message.
+     *
+     * @param parameters Any parameters required when processing the message.
+     */
+    void processMessage(ServiceParameters parameters);
+}
