--- src/main/java/uk/gov/companieshouse/company/appointments/InvalidMessageRouter.java
@@ -1,40 +1,0 @@
-package uk.gov.companieshouse.company.appointments;
-
-import org.apache.kafka.clients.producer.ProducerInterceptor;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
-
-import java.util.Map;
-
-/**
- * Routes a message to the invalid letter topic if a non-retryable error has been thrown during message processing.
- */
-public class InvalidMessageRouter implements ProducerInterceptor<String, String> {
-
-    private MessageFlags messageFlags;
-    private String invalidMessageTopic;
-
-    @Override
-    public ProducerRecord<String, String> onSend(ProducerRecord<String, String> producerRecord) {
-        if (messageFlags.isRetryable()) {
-            messageFlags.destroy();
-            return producerRecord;
-        } else {
-            return new ProducerRecord<>(this.invalidMessageTopic, producerRecord.key(), producerRecord.value());
-        }
-    }
-
-    @Override
-    public void onAcknowledgement(RecordMetadata metadata, Exception exception) {
-    }
-
-    @Override
-    public void close() {
-    }
-
-    @Override
-    public void configure(Map<String, ?> configs) {
-        this.messageFlags = (MessageFlags) configs.get("message.flags");
-        this.invalidMessageTopic = (String) configs.get("invalid.message.topic");
-    }
-}
