--- src/main/java/uk/gov/companieshouse/company/appointments/NullService.java
@@ -1,15 +1,0 @@
-package uk.gov.companieshouse.company.appointments;
-
-import org.springframework.stereotype.Component;
-
-/**
- * The default service.
- */
-@Component
-class NullService implements Service {
-
-    @Override
-    public void processMessage(ServiceParameters parameters) {
-        throw new NonRetryableException("Unable to handle message");
-    }
-}
