--- src/main/java/uk/gov/companieshouse/company/appointments/RetryableException.java
@@ -1,15 +1,0 @@
-package uk.gov.companieshouse.company.appointments;
-
-/**
- * A recoverable exception has occurred e.g. due to a service that is temporarily unavailable.
- */
-public class RetryableException extends RuntimeException {
-
-    public RetryableException(String message) {
-        super(message);
-    }
-
-    public RetryableException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
