+++ src/test/java/uk/gov/companieshouse/appointments/subdelta/TestUtils.java
@@ -1,0 +1,18 @@
+package uk.gov.companieshouse.appointments.subdelta;
+
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.clients.consumer.ConsumerRecords;
+
+public final class TestUtils {
+
+    private TestUtils(){
+    }
+
+    public static int noOfRecordsForTopic(ConsumerRecords<?, ?> records, String topic) {
+        int count = 0;
+        for (ConsumerRecord<?, ?> ignored : records.records(topic)) {
+            count++;
+        }
+        return count;
+    }
+}
