--- src/test/java/uk/gov/companieshouse/company/appointments/ErrorConsumerPositiveTest.java
@@ -1,73 +1,0 @@
-package uk.gov.companieshouse.company.appointments;
-
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
-import org.springframework.kafka.test.EmbeddedKafkaBroker;
-import org.springframework.kafka.test.context.EmbeddedKafka;
-import org.springframework.kafka.test.utils.KafkaTestUtils;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.TestPropertySource;
-
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.Mockito.verify;
-
-@SpringBootTest(classes = Application.class)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-@EmbeddedKafka(
-        topics = {"echo", "echo-echo-consumer-retry", "echo-echo-consumer-error", "echo-echo-consumer-invalid"},
-        controlledShutdown = true,
-        partitions = 1
-)
-@TestPropertySource(locations = "classpath:application-test_error_positive.yml")
-@Import(TestConfig.class)
-@ActiveProfiles("test_error_positive")
-class ErrorConsumerPositiveTest {
-
-    @Autowired
-    private EmbeddedKafkaBroker embeddedKafkaBroker;
-
-    @Autowired
-    private KafkaConsumer<String, String> testConsumer;
-
-    @Autowired
-    private KafkaProducer<String, String> testProducer;
-
-    @Autowired
-    private CountDownLatch latch;
-
-    @MockBean
-    private Service service;
-
-    @Test
-    void testConsumeFromErrorTopic() throws InterruptedException {
-        //given
-        embeddedKafkaBroker.consumeFromAllEmbeddedTopics(testConsumer);
-
-        //when
-        testProducer.send(new ProducerRecord<>("echo-echo-consumer-error", 0, System.currentTimeMillis(), "key", "value"));
-        if (!latch.await(30L, TimeUnit.SECONDS)) {
-            fail("Timed out waiting for latch");
-        }
-
-        //then
-        ConsumerRecords<String, String> consumerRecords = KafkaTestUtils.getRecords(testConsumer, 10000L, 1);
-        assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, "echo"), is(0));
-        assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, "echo-echo-consumer-retry"), is(0));
-        assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, "echo-echo-consumer-error"), is(1));
-        assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, "echo-echo-consumer-invalid"), is(0));
-        verify(service).processMessage(new ServiceParameters("value"));
-    }
-}
