--- src/test/java/uk/gov/companieshouse/company/appointments/TestUtils.java
@@ -1,18 +1,0 @@
-package uk.gov.companieshouse.company.appointments;
-
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-
-public final class TestUtils {
-
-    private TestUtils(){
-    }
-
-    public static int noOfRecordsForTopic(ConsumerRecords<?, ?> records, String topic) {
-        int count = 0;
-        for (ConsumerRecord<?, ?> ignored : records.records(topic)) {
-            count++;
-        }
-        return count;
-    }
-}
